svn: revision classes renamed
Now, the interface is called SvnRevision and the leaf types are called
SvnRepoRevision for repository revisions and SvnWorkingCopyRevision for
the "revision" of a working copy.
